Simplify selectAccount()
This uses jquery instead of getElementById and removes the hardcoded
list of pathnames. Previously, this list was used to decide what field
to focus,  but it seems there is always just one of these fields
present, so putting them both in a single jquery selector should let
jquery figure out which on is present.

This fixes "Rabobank site location not recognised" error popups when the
url didn't perfectly match (in particular /klanten/ instead of
/klanten).