Modify the "value" property, not attribute of input fields
The "value" attribute is the attribute in the DOM / from the HTML, which
sets the initial value of an input field. When the field was
not modified yet, it also sets the current value, but when the field was
modified, changing the "value" attribute not longer has any effect.

The current value of a field is stored in its "value" property (as
opposed to attribute), so modify that one instead.