Fix ideal page
This updates handling to support the updated ideal payment page as well.
Previously changes were only tested using the login and regular sending
pages, but it turns the ideal page was significantly changed as well.

Since on the ideal page, the account list could not be prepended to any
container (it should end up between two existing elements), the code is
changed to insert it before an existing element, instead of prepending
it to a container. This also works for the other pages (by using the
proper selector).