Fix login with the Rabo Scanner
It previously looked like the Rabo Scanner page worked, but now that I
actually have a scanner, it turns out it didn't yet. When entering an
account and card number with the scanner, the form needs to be submitted
so the server can generate the color code.

The code is changed to do this form submit, and some code is added to
prevent changing the number back to the default when the page is
reloaded after the form submit.