Bump jquery dependency
Jquery no longer seemed to work, bumping it this newer version fixes it.
It might not be the actual version that fixes things, but just that the
url is changed, causing Greasemonkey to download the jquery file (which
apparently happens when the script is installed or changed). But having
a recent jquery version won't hurt in any case.