Re-add support for sign pages
In commit 117df34 (Update to new Rabobank login page), the code was
changed to match the new login pages, under the assumption that the sign
pages would change similarly. However, 20 months later, this has not
happened yet, so this re-adds some of the selectors and code to support
the old-style sign pages again.