Big refactor to support "new" Rabobank sign page
This page was a bit tricky, since it uses a "shadow root" (dom root
inside a dom element) that prevents normal jquery selectors from looking
inside.

To handle this, all jquery selectors need to explicitly specify their
scope (and pass the shadow root or an element inside that). For all the
stuff that happens inside the selection panel, this is a matter of
passing `selectionPanel`. For other elements, this is more tricky, so
now the code is changed to find an store *all* relevant elements before
initialization, and then the rest of the code can just use those stored
elements. This has the added advantage that it is now a lot more clear
which selector is intended for which (generation of) page.

Additionally, the content of this new sign page is loaded
asynchronously, so we will have to try a few times to see if it is
loaded yet. There might be some dom complete event on the shadow root
that could be used here, but just a few retries also seems to work fine.

This also removes some options that were probably not really used, now
the script always removes the original "remember me" checkbox and no
longer allows using anything other than the first account as the
default.

Finally, the new sign page has no form to submit (like the login page)
to generate a challenge, but instead does asynchronously whenever the
account and pass number have been entered. To trigger this, input events
are faked after setting the values. It turns out this also works on the
login page, so the old form submit code has been removed.